+++ c/tskit/core.c
@@ -352,6 +352,10 @@ tsk_strerror_internal(int err)
         case TSK_ERR_NONBINARY_MUTATIONS_UNSUPPORTED:
             ret = "Only binary mutations are supported for this operation";
             break;
+        case TSK_ERR_UNION_NOT_SUPPORTED:
+            ret = "Union is not supported for cases where there is non-shared"
+                  "history older than the shared history of the two Table Collections";
+            break;
 
         /* Stats errors */
         case TSK_ERR_BAD_NUM_WINDOWS:
@@ -441,6 +445,16 @@ tsk_strerror_internal(int err)
         case TSK_ERR_TOO_MANY_VALUES:
             ret = "Too many values to compress";
             break;
+
+        /* Union errors */
+        case TSK_ERR_UNION_BAD_MAP:
+            ret = "Node map contains an entry of a node not present in this table "
+                  "collection.";
+            break;
+        case TSK_ERR_UNION_DIFF_HISTORIES:
+            // histories could be equivalent, because subset does not reorder
+            // edges (if not sorted) or mutations.
+            ret = "Shared portions of the tree sequences are not equal.";
     }
     return ret;
 }